requirements.txt
numpy==1.23.5
pyYAML
opencv-python
uncertainty_toolbox
brisque
tensorflow==2.10.0
tensorflow-probability==0.17.0
imagehash
seaborn
tensorflow-object-detection-api==0.1.1
ijson
albumentations
tensorflow_addons
pycocotools
tensorflow_hub==0.12.0
neural-structured-learning
hebo
optuna
jupyterlab
torch==1.12.1